Force token_type_ids to int type

+++ bertblocks/modeling/embedding.py
@@ -50,7 +50,7 @@ class TokenTypeEmbedding(nn.Module):
         """
         shape = x.shape[:2] if len(x.shape) == 3 else x.shape[:1]
         token_type_ids = token_type_ids if token_type_ids is not None else torch.zeros(shape, device=x.device)
-        return x + self.embd(token_type_ids)
+        return x + self.embd(token_type_ids.to(torch.int32))
 
 
 class TokenEmbedding(nn.Module):
